Turn SettingsActivity into SettingsFragment
Also converted over applicable tests. Some tests don't make much sense to move over as of right now (in my opinion), as we're not planning on keeping any of these Activities as they are.
