+++ app/src/main/java/com/nytimes/android/external/register/di/ActivityComponent.kt
@@ -2,7 +2,7 @@ package com.nytimes.android.external.register.di
 
 import com.nytimes.android.external.register.BuyActivity
 import com.nytimes.android.external.register.MainActivity
-import com.nytimes.android.external.register.SettingsActivity
+import com.nytimes.android.external.register.SettingsFragment
 
 import dagger.Subcomponent
 
@@ -11,5 +11,5 @@ import dagger.Subcomponent
 interface ActivityComponent {
     fun inject(activity: MainActivity)
     fun inject(activity: BuyActivity)
-    fun inject(activity: SettingsActivity)
+    fun inject(activity: SettingsFragment)
 }
